added authoriztion to tasks

+++ src/tasks/tasks.service.ts
@@ -7,6 +7,7 @@ import { Repository } from 'typeorm';
 import { CreatTaskDto } from './dto/create-task.dto';
 import { TaskStatus } from './task-statu.enum';
 import { GetTasksFilterDto } from './dto/get-task-filter.dto';
+import { UserEntity } from 'src/auth/user.entity';
 
 @Injectable()
 export class TasksService {
@@ -15,10 +16,12 @@ export class TasksService {
     private taskRepository: Repository<TaskEntity>,
   ) {}
 
-  async getTasks(filterDto: GetTasksFilterDto) {
+  async getTasks(filterDto: GetTasksFilterDto, user: UserEntity) {
     const { status, search } = filterDto;
 
     const query = this.taskRepository.createQueryBuilder('task');
+
+    query.where('task.userId = :userId', { userId: user.id });
     if (status) {
       query.andWhere('task.status = :status', { status });
     }
@@ -34,8 +37,10 @@ export class TasksService {
     return tasks;
   }
 
-  async getTaskById(id: number): Promise<TaskEntity> {
-    const found = await this.taskRepository.findOneBy({ id });
+  async getTaskById(id: number, user: UserEntity): Promise<TaskEntity> {
+    const found = await this.taskRepository.findOne({
+      where: { id, userId: user.id },
+    });
 
     if (!found) {
       throw new NotFoundException(`Task with ID ${id} not found`);
@@ -43,63 +48,40 @@ export class TasksService {
     return found;
   }
 
-  async deleteTask(id: number): Promise<void> {
-    const found = await this.taskRepository.findOneBy({ id });
-    if (!found) {
-      throw new NotFoundException(`Task with ID ${id} not found`);
+  async deleteTask(id: number, user: UserEntity): Promise<void> {
+    const result = await this.taskRepository.delete({ id, userId: user.id });
+
+    if (result.affected === 0) {
+      throw new NotFoundException(`Task with ID "${id}" does not exist`);
     }
-    await this.taskRepository.delete({ id });
   }
 
-  async createTask(createTaskDto: CreatTaskDto): Promise<TaskEntity> {
+  async createTask(
+    createTaskDto: CreatTaskDto,
+    user: UserEntity,
+  ): Promise<TaskEntity> {
     const { title, description } = createTaskDto;
 
     const task = new TaskEntity();
     task.title = title;
     task.description = description;
     task.status = TaskStatus.OPEN;
+    task.user = user;
     await task.save();
+
+    delete task.user;
 
     return task;
   }
 
-  async updateTask(id: number, status: TaskStatus): Promise<TaskEntity> {
-    const task = await this.getTaskById(id);
+  async updateTask(
+    id: number,
+    status: TaskStatus,
+    user: UserEntity,
+  ): Promise<TaskEntity> {
+    const task = await this.getTaskById(id, user);
     task.status = status;
     await task.save();
     return task;
   }
-
-  // getTaskById(id: string): Task {
-  //   const found = this.tasks.find((task) => task.id === id);
-
-  //   if (!found) {
-  //     throw new NotFoundException(`Task with ID ${id} not found`);
-  //   }
-  //   return found;
-  // }
-
-  // createTask(createTaskDto: CreatTaskDto): Task {
-  //   const { title, description } = createTaskDto;
-  //   const task: Task = {
-  //     id: uuid(),
-  //     title,
-  //     description,
-  //     status: TaskStatus.OPEN,
-  //   };
-
-  //   this.tasks.push(task);
-  //   return task;
-  // }
-
-  // deleteTask(id: string): void {
-  //   const found = this.getTaskById(id);
-  //   this.tasks = this.tasks.filter((task) => task.id !== found.id);
-  // }
-
-  // updateTaskStatus(id: string, status: TaskStatus): Task {
-  //   const task = this.getTaskById(id);
-  //   task.status = status;
-  //   return task;
-  // }
 }
